% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_changeNames.R
\name{changeNames}
\alias{changeNames}
\title{Rename (some) in a object and return it}
\usage{
changeNames(x, i = 1:length(x), nm)
}
\arguments{
\item{x}{Any object with elements that can be named (atomic vector, list, data.frame, ...).}

\item{i}{Vector of indices (integers or logical) of elements in x to rename.}

\item{nm}{}
}
\value{
The same object as x but with names of elements i changed.
}
\description{
Just like `setNames` but you can provide indices to set/change only selected names.
}
