% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cohend.R
\name{cohend}
\alias{cohend}
\alias{sd_pool}
\title{Compute pooled standard deviation and/or Cohen's d effect size}
\usage{
cohend(x, y)

sd_pool(x, y, n1 = NA, n2 = NA)
}
\arguments{
\item{x}{A numerical vector or (for `sd_pool`) an SD value.}

\item{y}{A numerical vector or (for `sd_pool`) an SD value.}

\item{n1}{If x is an SD, value, provide the size of sample x as a numerical vector of length 1.}

\item{n2}{If x is an SD, value, provide the size of sample y as a numerical vector of length 1.}
}
\value{
Both functions return a umerical vector of length 1 giving the pooled SD or Cohen's d effect size.
}
\description{
Compute pooled standard deviation and/or Cohen's d effect size
}
