% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_db_con.R
\name{db_con}
\alias{db_con}
\title{Connect to IBEs databases}
\usage{
db_con(
  user,
  db = c("all", "itembank", "checkauswertung", "mutterdatei", "tracking",
    "check_data_revisions")
)
}
\arguments{
\item{user}{Username to log on to DB server with.}

\item{db}{Character: Which database to connect to. If several or "all" are chosen, the output will be a list of connections.}
}
\value{
A connection object of list of connection objects.
}
\description{
Wrapper around RMariaDB::dbConnect to connect to itembank, checkauswertung (sessions / results), and check_data_revisions (mutterdatei). An active SSH tunnel is required!
}
