% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_empty_sql_table.R
\name{df_to_empty_sql}
\alias{df_to_empty_sql}
\title{Create empty sql table based on r data frame data types}
\usage{
df_to_empty_sql(df, conn, table_name)
}
\arguments{
\item{df}{Data frame with defined column data types}

\item{conn}{Established connection to database}

\item{table_name}{Name of table in database}
}
\value{
SQL query that was used for creating empty table
}
\description{
Create empty sql table based on r data frame data types
}
