% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_domain_name2mx.R
\name{domain_name_matrix}
\alias{domain_name_matrix}
\alias{domain_name2mx}
\alias{domain_mx2name}
\title{Translate between domain names and matrix codes}
\usage{
domain_name2mx(x, stop.when.unknown = FALSE)

domain_mx2name(x, stop.when.unknown = FALSE)
}
\arguments{
\item{x}{A character vector of domain names or codes.}

\item{stop.when.unknown}{Logical, whether to stop with an error when one or several values of x are invalid. If `FALSE` (default), warnings are given for unknown domains or competence areas.}
}
\value{
A character vector of the same length as x, potentially including new `NA`s.
}
\description{
`domain_name2mx` translates domain names (e.g. "dles", "natw") to short matrix codes according to the Lehrplan 21 (e.g. "D.2", "NT").
`domain_mx2name` translates in the opposite direction. Domain name "natw" translates to "NT" and "math" to "MA".
}
