% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_duplicated2.R
\name{duplicated2}
\alias{duplicated2}
\title{Check for duplicates from both ends}
\usage{
duplicated2(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or `NULL`}

\item{incomparables}{a vector of values that cannot be compared. `FALSE` is a special value, meaning that all values can be compared, and may be the only value accepted for methods other than the default. It will be coerced internally to the same type as `x`.}

\item{...}{arguments for particular methods}
}
\value{
For a vector input, a logical vector of the same length as x. For a data frame, a logical vector with one element for each row. For a matrix or array, and when MARGIN = 0, a logical array with the same dimensions and dimnames.
}
\description{
The base function `duplicated(x)` returns a logical vector where every duplicated element of `x` is `TRUE`. In order to get the first element, too, we have to use `duplicated(x) | duplicated(x, fromLast=TRUE)`. `duplicated2()` is a wrapper around `duplicated()` that makes this easyier.
}
\details{
`duplicated()` has several methods for different types of `x`. These should work identically for `duplicated2()` because it does nothing but pass the arguments.
}
\examples{
a <- c(1, 2, 3, 3, 4)
duplicated(a)
duplicated(a, fromLast=TRUE)
duplicated2(a)


}
\seealso{
[duplicated()] which this function wraps.
}
