% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_find_trans_via_quantiles.R
\name{find_trans_via_quantiles}
\alias{find_trans_via_quantiles}
\title{Find linear transformation via quantiles}
\usage{
find_trans_via_quantiles(
  x = NULL,
  y = NULL,
  Qx = NULL,
  Qy = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{x}{Raw values of distribution *X*}

\item{y}{Raw values of distribution *Y*}

\item{Qx}{Quantiles of distribution *X* (provide *either* `x` or `Qx`).}

\item{Qy}{Quantiles of distribution *X* (provide *either* `x` or `Qx`).}

\item{probs}{Probabilities of quantiles to be computed. Only needed if `Qx` or `Qy` (or both) not provided.
Must be a numeric vector of length >= 2 and values between 0 and 1. Defaults to `c(.05, .25, .50, .75, .95)`.}
}
\value{
A named vector with the elements `constant` and `multiplier`.
}
\description{
Find the contant and multiplier to linearly transform *X* into *Y* based on quantiles of their respective distributions.
  Basically a wrapper around `lm()` where the quantiles of *Y* are regressed onto *Y*.
  For both variables, provide either raw values or given quantiles.
}
