% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_get_abil_output.R
\name{get_abil_output_file}
\alias{get_abil_output_file}
\title{Get ability-output files (input into TAM script) from IBE server}
\usage{
get_abil_output_file(
  check,
  year,
  domain,
  selection = c("date", "name"),
  check_type = substr(toupper(check), 1, 1),
  spec_regex = NULL,
  read_data = TRUE,
  avoid.nz = TRUE
)
}
\arguments{
\item{check}{Check type, i.e. P3, P5, S2 or S3}

\item{year}{Year of check}

\item{domain}{Scale/test, e.g. "dles". "math" is *not* vaild, but "mzuv", "mfur", & "mgfd" are valid.}

\item{selection}{If several files are found, choose either the newest (date) or the last among sorted file names.}

\item{check_type}{`"P"` for primary school or `"S"` for secondary school, by default consistent with `check`}

\item{spec_regex}{The function constructs a regular expression to find the correct file(s). Provide your own regular expression here, to overrule.}

\item{read_data}{If TRUE (default), use readRDS and return file content; if FALSE, return file path.}

\item{avoid.nz}{Avoids Nachzügler files by excluding file names that contain "NZ"}
}
\value{
If `read_data` is `TRUE`, the content of the Rds-file as a data frame, else the full path to the Rds-file.
}
\description{
Get ability-output files (input into TAM script) from IBE server
}
