% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_hid2canton.R
\name{hid2canton}
\alias{hid2canton}
\title{Gets canton abbreviation or name from HID as used in Checks/Mindsteps-system}
\usage{
hid2canton(hid, full.name = FALSE)
}
\arguments{
\item{hid}{A vector of HIDs (canton, school, class, or student)}

\item{full.name}{Return full name of canton or abbreviation (e.g., ZH)? Default is abbreviation}
}
\value{
A character vector of same length as `hid`.
}
\description{
Gets canton abbreviation or name from HID as used in Checks/Mindsteps-system
}
