% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_id_srv_path.R
\name{id_srv_path}
\alias{id_srv_path}
\title{Test if ID_Daten connection is available via id-srv or IP-address}
\usage{
id_srv_path(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to show a message of the result.}
}
\value{
A character vector of length 1: The path to ID_Daten if available, otherwise NA_character_.
}
\description{
Test if ID_Daten connection is available via id-srv or IP-address
}
\details{
Checks for ID_Daten via id-srv and the IP-adress. If both are available, id-srv is preferred.
}
