% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lintrans.R
\name{lintrans_distr}
\alias{lintrans_distr}
\title{Get linear transformation based on distribution parameters (M, SD) or quantiles}
\usage{
lintrans_distr(x, y, use_prop = 1)
}
\arguments{
\item{x}{Numeric vector of the original distribution.}

\item{y}{Numeric vector of the target distribution.}

\item{use_prop}{Numeric vector of value `0 < use_prop <= 1` Proportion of data used in each distribution vector. If < 1, equal percentages of data are removed from both ends of the distribution.}

\item{robust}{Whether to use a robust regression method instead of ordinary least squares regression. Requires the package `robustbase`.}
}
\value{
A list of two numeric vectors of length 1 each: constant and multiplier.
}
\description{
Get linear transformation coefficients (constant and multiplier) to transform from one distribution to the other.
`lintrans_distr` uses the mean and standard deviation. `lintrans_regr` uses the coefficients of a linear regression. The latter method requires that the data sets are paired.
}
