% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_makeNA.R
\name{makeNA}
\alias{makeNA}
\title{Make NAs of given data type or matching given vector}
\usage{
makeNA(x = NULL, type = NULL, length = 1)
}
\arguments{
\item{x}{An atomic (!) vector of which the data type and length is used. Omit this if you provide `type`.}

\item{type}{A data type to use. Omit this if you provide `x`.}

\item{length}{The length of the output, defaults to 1 if `x` is not given.}
}
\value{
An atomic vector of NA values matching type and length of `x` or of type `type` and length `length`.
}
\description{
Generate a vector of NA values of a specific data type. Either provide an example vector or specify the type.
This is useful for avoiding errors in picky functions (e.g. `dplyr::if_else`).
Remember that "numeric" is not a type (use "real" or "integer" instead), it is turned into "real" instad, with a warning.
}
\examples{
makeNA(1:10)
makeNA(1L:10L)
makeNA(type = "character", length = 5)

df <- data.frame(a = letters,
                 x = 1:26,
                 y = rep(c(T, F), 13))
# results in an error as NA is of wrong type
for (i in 1:2) dplyr::if_else(df$y, df[[i]], NA)
# no error as NA value is adapted to type
for (i in 1:2) dplyr::if_else(df$y, df[[i]], makeNA(df[[i]]))

}
