% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_naFill.R
\name{na.fill}
\alias{na.fill}
\alias{na.fill.t}
\alias{na.fill.f}
\title{Fill NAs with replacement values}
\usage{
na.fill(x, y, warn = TRUE)

na.fill.t(x, warn = FALSE)

na.fill.f(x, warn = FALSE)
}
\arguments{
\item{x}{A vector that may include `NA`s}

\item{y}{A vector of values used to replace `NA`s in `x` with}

\item{warn}{Whether to give warnings about irregularities}
}
\value{
Vector `x` with `NA`s replaced by values from `y`.
}
\description{
Avoid NAs by filling them with given values. The length of the original vector is retained.
`na.fill.t` and `na.fill.f` are wrappers around `na.fill` to fill NAs in logical vectors with `TRUE` or `FALSE`, respectively.
}
\details{
If the length of `x` is not a multiple of `y`s length only the first value in `y` is used, potentially with a warning. 
`na.fill.t` and `na.fill.f` coerce `x` to logical, potentially with a warning.
}
\examples{
# usage to replace NAs
bla <- c(1:5, NA_real_, 7:10)
na.fill(bla, 100)
na.fill(bla, 101:110)

# usage in indexing
i <- c(T, T, F, F, T, NA_logical_, F, F, T, T)
try(LETTERS[i]) # results in an error, as NAs are not allowed for indices
LETTERS[na.omit(i)]
LETTERS[na.fill.t(i)]
LETTERS[na.fill.f(i)]

}
