% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_nra_cells.R
\name{nra_cells}
\alias{nra_cells}
\title{Create non-response adjustment cells}
\usage{
nra_cells(df, nResp = 15, nra.sep, nra.sort, respVar, respValue, clustVar)
}
\arguments{
\item{df}{A data frame with one case per row, where respondents can be identified using a participation status variable.}

\item{nResp}{The minimum number of respondents that an nra cell must contain; default is 15.}

\item{nra.sep}{character string or vector with variable names based on which separate cell formations are created (e.g. regions, gender, etc.).}

\item{nra.sort}{character string or vector with variable names on the basis of which the data is sorted so that cases which are similar with respect to this variables end up in the same nra cells.}

\item{respVar}{character string with name of variable that defines respondents}

\item{respValue}{numeric value of respVar defining respondents}

\item{clustVar}{character string with name of clustering variable; clusters are not split across cells (e.g. schools)}
}
\description{
Create non-response adjustment cells
}
