% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_nra_weight_calculation.R
\name{nra_weights}
\alias{nra_weights}
\title{Calculates non-response adjusted weights}
\usage{
nra_weights(
  df,
  cellVar,
  weightVar,
  nraWeightVar,
  statusVar,
  nonRespValue,
  respValue
)
}
\arguments{
\item{df}{Weighted data frame with nra cells defined; use nra_cells() for cell definition}

\item{cellVar}{Name of nra cell variable}

\item{weightVar}{Name of weight variable that needs adjustments}

\item{nraWeightVar}{Name of output variable}

\item{statusVar}{Name of integer variable containing participation codes, where one code represents respondents}

\item{nonRespValue}{Integer or integer vector containing statusVar values that represent non-responding cases for which responder weights are adjusted}

\item{respValue}{Integer for respVar value that represents responding cases}
}
\value{
Inout data frame with additional column nraWeightVar containing adjusted weights
}
\description{
Calculates non-response adjusted weights
}
