% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_ICC.R
\name{plot_ICC}
\alias{plot_ICC}
\title{Plot Item Characteristic Curve(s)}
\usage{
plot_ICC(diff, disc, xlim = -5:5)
}
\arguments{
\item{diff}{Difficulty Parameter}

\item{disc}{Discrimination Parameter}

\item{xlim}{Theta display range}
}
\value{
ggplot object
}
\description{
Plot Item Characteristic Curve(s)
}
