% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_IIC.R
\name{plot_IIC}
\alias{plot_IIC}
\title{Plot Item Information Curve(s)}
\usage{
plot_IIC(b, a, xlim = -5:5)
}
\arguments{
\item{b}{difficulty parameter}

\item{a}{discrimination parameter}

\item{xlim}{Theta display range}
}
\value{
ggplot object
}
\description{
Plot Item Information Curve(s)
}
\examples{
plot_IIC(-1:1, c(.5, 1, 1.5))

}
