% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_QPDF.R
\name{QPDF}
\alias{QPDF}
\title{Encrypt PDF-files with QPDF}
\usage{
QPDF(path, input, output, pw, qpdf.path = NULL, overwrite = FALSE)
}
\arguments{
\item{path}{The path to directory containing the input PDF document.}

\item{input}{Name of the input PDF document as a string.}

\item{output}{Name of the output PDF document as a string.}

\item{pw}{Passwort as a string. Beware of escape characters!}

\item{qpdf.path}{The path to the qpdf.exe file. If NULL and the user is "bwolf", the location where BW has his exe. Other user's paths locations may be added in the future.}

\item{overwrite}{Logical. Should the output file be overwritten if it already exists? Results in an error if `FALSE` and the file exists.}
}
\value{
Returns -- invisibly -- a 0 if QPDF ran successfully or else the error code (usually 127).
}
\description{
This is wrapper around the qpdf.exe shell command for Windows.
It will not work on other systems and requires that the QPDF program is available in a writable directory on the system.
QPDF can be downloaded from https://sourceforge.net/projects/qpdf/files/qpdf/10.3.2/ .
}
\details{
The function checks the availablility of the input file and output file. A success or error message is given upon exit with
}
\examples{
QPDF("C:/Users/bwolf/Documents/", input = "mySecretDoc.pdf", output = "mySecretDoc_encrypted.pdf", pw = "123456")

}
