% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_round.R
\name{Round}
\alias{Round}
\title{Conventional rounding}
\usage{
Round(x, digits = 0)
}
\arguments{
\item{x}{A vector of real numbers}

\item{digits}{Decimal places to preserve}
}
\value{
A vector of numbers rounded to the indicated decimal places
}
\description{
Conventional rounding
}
