% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_pps.R
\name{sample_PPS}
\alias{sample_PPS}
\title{Select PPS-Sample of clusters}
\usage{
sample_PPS(df, mos, tcs, n)
}
\arguments{
\item{df}{Stratified data frame with clusters and corresponding MOS (measures of size)}

\item{mos}{Column name containing MOS (measure of size)}

\item{tcs}{Target cluster size}

\item{n}{Total number of elements (in clusters) to sample}
}
\value{
Input data frame is completed with columns containing sampling information. Selected schools are flagged by sc_smp_selected.
}
\description{
Select PPS-Sample of clusters
}
