% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_mixed_design.R
\name{sample_PPS2}
\alias{sample_PPS2}
\title{Mixed design cluster sampling with 3 strata: Small clusters (simple/systematic random sample), Medium clusters (PPS) and Large clusters (full census).}
\usage{
sample_PPS2(df, enrVar, tcs, n)
}
\arguments{
\item{df}{Stratified data frame with clusters and (estimated) elements per cluster}

\item{enrVar}{Character string with name of column containing student enrollment number}

\item{tcs}{Numeric target cluster size}

\item{n}{Total number of elements to sample}
}
\value{
Input data frame is completed with columns containing sampling information. Selected schools are flagged by sc_smp_selected.
}
\description{
Mixed design cluster sampling with 3 strata: Small clusters (simple/systematic random sample), Medium clusters (PPS) and Large clusters (full census).
}
