% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_serpentine_sort.R
\name{serpentine_sort}
\alias{serpentine_sort}
\title{Sorts data frame by multiple (stratification) variables in a serpentine manner}
\usage{
serpentine_sort(data, implstrat)
}
\arguments{
\item{data}{data frame in need of some serpentine sorting}

\item{implstrat}{character vector containing sorting variables}
}
\value{
sorted input data frame
}
\description{
Sorts data frame by multiple (stratification) variables in a serpentine manner
}
