% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_source_to_list.R
\name{source_to_list}
\alias{source_to_list}
\title{Source a file and store store objects in a list}
\usage{
source_to_list(file, only.functions = TRUE, ...)
}
\arguments{
\item{file}{Path to file to be sourced.}

\item{only.functions}{Whether only functions or also other objects generated by the code should be returned (defaults to `TRUE`).}

\item{...}{Other arguments to be passed to `source()`.}
}
\value{
A list of all objects generated by the sourced code. If `only.functions = TRUE`, then other objects than functions are removed.
}
\description{
This is a wrapper around `source`. It is mainly useful to produce a list of functions when several functions are defined in one file.
}
