% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_small_school_analysis.R
\name{ssa}
\alias{ssa}
\title{Small school analysis for PPS school samples}
\usage{
ssa(df, enrVar, tcs, n)
}
\arguments{
\item{df}{Data frame containing school frame of ONE explicit stratum}

\item{enrVar}{Character string with column name containing student enrollment numbers}

\item{tcs}{Numeric target cluster size}

\item{n}{Numeric amount of students to sample in stratum}
}
\value{
Table with number of schools to sample and info on undersampling of very small schools
}
\description{
Small school analysis for PPS school samples
}
