% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_db_table_exists.R
\name{table_exists}
\alias{table_exists}
\title{Check if table with particular name exists in database}
\usage{
table_exists(conn, table_name, database_name)
}
\arguments{
\item{conn}{Established connection to database}

\item{table_name}{Name of table in database}

\item{database_name}{Name of database}
}
\value{
Logical - TRUE if database table exists
}
\description{
Check if table with particular name exists in database
}
