% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_tageskombi.R
\name{tageskombi}
\alias{tageskombi}
\title{Makes evenly distributed group assignments for a given number of persons and days. Used for planning double-corrections in writing tests.}
\usage{
tageskombi(npers, ntage, maxzeit = 10, seed = 123, raters = NULL)
}
\arguments{
\item{npers}{number of persons (raters)}

\item{ntage}{number of days}

\item{maxzeit}{maximum time (seconds) to use for iterative process}

\item{seed}{initial seed to use. The search loop for valid combinations increments the seed by 1 after each iteration.}
}
\value{
a data frame with `npers` rows, a column `Gruppe` for group assignment (A and B), and `ntage` columns for days
}
\description{
Makes evenly distributed group assignments for a given number of persons and days. Used for planning double-corrections in writing tests.
}
