% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_unify_coltypes.R
\name{unify_coltypes}
\alias{unify_coltypes}
\title{Unify the types of variables across a list of data frames}
\usage{
unify_coltypes(df.list, verbose = TRUE)
}
\arguments{
\item{df.list}{A list of data frames}

\item{verbose}{Logical, whether to give a message for every column that is changed.}
}
\value{
The same list of data frames but colums that appear in several of them are coerced to the same type. 
The order of columns may still differ between data frames. The list can be submitted to `dplyr::bind_rows()`.
}
\description{
Coerces columns that appear in several data frames to the same type.
This is useful to do before stacking the data frames with `dplyr::bind_rows()`.
The order of priority of types is as follows: character > double > integer > logical
For instance, if any of the same-named columns is of type "character", then all others will be coerced to character.
Columns that appear in only one of the data frames or already have the same type across all data frames remain unchanged.
}
