% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_write_to_db.R
\name{write_df_to_db}
\alias{write_df_to_db}
\title{Saves df as temporary table and loads its content into an existing database table (OVERWRITE!) - use df_to_empty_sql if necessary}
\usage{
write_df_to_db(df, conn_name, table_name)
}
\arguments{
\item{df}{Data frame that needs to be written into a database}

\item{conn_name}{Established connection to database}

\item{table_name}{Name of table in database}
}
\value{
SQL query that was used for writing table
}
\description{
Saves df as temporary table and loads its content into an existing database table (OVERWRITE!) - use df_to_empty_sql if necessary
}
