% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_itembank_interactions.R
\name{connect_itembank}
\alias{connect_itembank}
\alias{connect_checkbank}
\alias{dbDataTypes}
\title{Custom functions for DB interactions}
\usage{
connect_itembank(user, pw = NULL)

connect_checkbank(user, pw = NULL)

dbDataTypes(con, tab, incl_length = TRUE, string_only = FALSE)
}
\arguments{
\item{user}{The username as a string.}

\item{pw}{The password as a string. If `NULL`, the default, a pop-up appears to ask for the password.}

\item{con}{A database connection object, as generated by `connect_itembank` or `dbConnect`.}

\item{tab}{The name of a table in the database, as a string.}

\item{incl_length}{Logical, whether to include the maximum character length of each column.}

\item{string_only}{Logical, whether to filter the output for columns of type `string` only.}
}
\value{
For `connect_itembank` and `connect_checkbank`, a database connection object, class `MariaDBConnection`. For `dbDataTypes` a data.frame with one row per column in the database table.
}
\description{
`connect_itembank` and `connect_checkbank` are wrappers around `RMariaDB::dbConnect` that include the host and port needed to access our DBs.
`dbDataTypes` gets the data storage types and - if chosen - the character lengths of each column.
}
