% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_make_luhn.R
\name{make_luhn}
\alias{make_luhn}
\title{Make Luhn checksums as needed for HIDs (including scan-IDs)}
\usage{
make_luhn(x, HID = TRUE, sep = ifelse(HID, "-0", ""))
}
\arguments{
\item{x}{A number or character string of numbers. All characters and special characters are ignored when calculating the check sum but preserved in the output.}

\item{HID}{Set `TRUE` for HID-formatted output.}

\item{sep}{Separator between the original input `x` and the checksum that is appended.}
}
\value{
A character string pasted from the original input `x`, the separator `sep`, and the checksum.
}
\description{
Make Luhn checksums as needed for HIDs (including scan-IDs)
}
\examples{
sapply(c(2, 6, 13, 17), function(nziff) make_luhn(substring("01-999-88-777-666", 1, nziff)))

}
