% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_ICC.R
\name{plot_ICC}
\alias{plot_ICC}
\title{Plot Item Characteristic Curve(s)}
\usage{
plot_ICC(b, a, xlim = -5:5)
}
\arguments{
\item{b}{difficulty parameter}

\item{a}{discrimination (slope) parameter}

\item{xlim}{theta display range, defaults to -5 to 5}
}
\value{
ggplot object
}
\description{
Plot Item Characteristic Curve(s)
}
\examples{
plot_ICC(rep(-1:1, 2), rep(1:2, each=3))

}
