% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_IIC.R
\name{plot_IIC}
\alias{plot_IIC}
\title{Plot Item Information Curve(s)}
\usage{
plot_IIC(diff, disc, xlim = -5:5)
}
\arguments{
\item{diff}{Difficulty Parameter}

\item{disc}{Discrimination Parameter}

\item{xlim}{Theta display range}
}
\value{
ggplot object
}
\description{
Plot Item Information Curve(s)
}
