% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_round.R
\name{Round}
\alias{Round}
\title{Conventional rounding}
\usage{
Round(x, digits = 0)
}
\arguments{
\item{x}{A vector of real numbers}

\item{digits}{Decimal places to preserve}
}
\value{
A vector of numbers rounded to the indicated decimal places
}
\description{
Rs `round` function rounds to the nearest even number, so that 1.5 becomes 2 but 0.5 becomes 0. In contrast, `Round` always rounds .5 upwards.
}
\examples{
a <- seq(-3, 3, by = .5)
round(a)
Round(a)

}
