% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampling_pps.R
\name{sample_PPS}
\alias{sample_PPS}
\title{Select PPS-Sample of clusters}
\usage{
sample_PPS(df, enrVar, tcs, n, undersampling = FALSE, idStart = 1)
}
\arguments{
\item{df}{Stratified data frame with clusters and (estimated) elements per cluster}

\item{enrVar}{Character string with name of column containing student enrollment number}

\item{tcs}{Numeric target cluster size}

\item{n}{Total number of clusters to sample}

\item{undersampling}{Logical: If TRUE, very small and tiny schools are undersampled by factor 2 and 4}

\item{idStart}{Numeric at which identificators of selected clusters start.}
}
\value{
Input data frame is completed with columns containing sampling information. Selected schools are flagged by sc_smp_selected.
}
\description{
Select PPS-Sample of clusters
}
