% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_source_to_list.R
\name{source_to_list}
\alias{source_to_list}
\title{Source and store objects as list}
\usage{
source_to_list(files, sorted = TRUE, only.functions = FALSE, ...)
}
\arguments{
\item{files}{The file or a list/vector of files passed to the `file` argument of `source`.}

\item{sorted}{Sort objects by name or keep in order of generation?}

\item{only.functions}{Whether only functions or also other objects generated by the code should be returned}

\item{...}{Further arguments passed to `source`. This must not contain the argument `local`!}
}
\value{
A named list of objects, possibly sorted by name, optionally excluding all non-functions.
}
\description{
Executes the code provided in a given connection (file or URL) or list of connections and stores all generated objects in a named list.
Useful, for instance, to get a list of functions defined in one or more external scripts.
}
\details{
If `files` is a vector (including lists) with more than 1 element, each of the elements is passed to `scource` consecutively.

According to the documentation of `base::source`, the `file` argument is: a connection or a character string giving the pathname of the file or URL to read from. "" indicates the connection `stdin()`.
}
